+++ bundles/org.eclipse.wst.xml.xpath2.processor/src/org/eclipse/wst/xml/xpath2/processor/internal/function/FnSecondsFromDuration.java
@@ -7,7 +7,8 @@
  *
  * Contributors:
  *     Andrea Bittau - initial API and implementation from the PsychoPath XPath 2.0
- *     Mukul Gandhi - bug 273760 - wrong namespace for functions and data types 
+ *     Mukul Gandhi - bug 273760 - wrong namespace for functions and data types
+ *     David CArver - bug 277774 - XSDecimal returning wrong values. 
  *******************************************************************************/
 
 package org.eclipse.wst.xml.xpath2.processor.internal.function;
@@ -18,6 +19,7 @@ import org.eclipse.wst.xml.xpath2.processor.ResultSequenceFactory;
 import org.eclipse.wst.xml.xpath2.processor.internal.*;
 import org.eclipse.wst.xml.xpath2.processor.internal.types.*;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 /**
@@ -77,7 +79,7 @@ public class FnSecondsFromDuration extends Function {
 		if (dtd.negative())
 			res *= -1;
 
-		rs.add(new XSDecimal(res));
+		rs.add(new XSDecimal(BigDecimal.valueOf(res)));
 
 		return rs;
 	}
